economics/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
numpy==1.24.3
numpy-financial==1.0.0
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
python-dotenv==1.0.0
marshmallow==3.20.1
werkzeug==2.3.7
gunicorn==21.2.0 